refactor: move sign up related classes to shared auth package
This commit moves the `SignUpRequest` and `SignUpResponse` classes to the `shared.
auth.sign_up` package to better organize the codebase and improve maintainability.

Additionally, the `SignInScreen` is updated to use the `OutlinedTextField` component for the password field, providing a more consistent look and feel with other input fields. The password field also now supports password visibility toggling.

Finally, the `SignInViewModel` is enhanced with a `checkProfile` function to determine if the user has already created a profile, and updates the navigation state accordingly.